+ 穿模fixed + 隔墙杀fixed
