+++ Assets/James folder/Scripts/GhostScript.cs
@@ -6,39 +6,23 @@ public class GhostScript : MonoBehaviour
 {
     public bool onDoor;
 	public GameObject targetSetter;
-	private AITargetScript ts;
+	
 	//public Animator GhostAnim;
 	public Animation GhostAnim;
-	public GameObject camControl;
+	
 
-	private void Start()
-	{
-		ts = targetSetter.GetComponent<AITargetScript>();
-	}
+	
 
     private void Update()
     {
 		GhostAnimation();
 	}
 
-    private void OnTriggerEnter(Collider other)
-	{
-		if (other.CompareTag("Player"))
-		{
-			print("collide with player");
-			if (ts.playerLightSource != null)
-			{
-				print("kill player");
-				targetSetter.SetActive(false); // disable target setter so ai won't move anymore
-				camControl.GetComponent<CamControl>().isDead = true;  //Player die here.
-				// implement player die here
-			}
-		}
-	}
+    
 
 	void GhostAnimation()
     {
-		if (this.transform.position == targetSetter.transform.position)
+		if (transform.position == targetSetter.transform.position)
 		{
 			GhostAnim.Play("idle");
 
@@ -46,5 +30,4 @@ public class GhostScript : MonoBehaviour
 		else
 			GhostAnim.Play("walk");
     }
-
 }
