+++ Assets/James folder/Scripts/KillPlayerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 86b4e18d26f6cd74b852f62c58b2fc06
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
